+++ apidoc/html/search/all_13.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['rc1_202025_2005_2006_0',['[1.0.1-rc1] - 2025-05-06',['../md_CHANGELOG.html#autotoc_md1',1,'']]],
-  ['representations_1',['Escaped string representations',['../index.html#autotoc_md28',1,'']]]
+  ['representations_0',['Escaped string representations',['../index.html#autotoc_md29',1,'']]]
 ];