+++ apidoc/html/search/all_5.js
@@ -1,10 +1,11 @@
 var searchData=
 [
-  ['accessing_20substructures_20and_20elements_0',['Accessing substructures and elements',['../index.html#autotoc_md21',1,'']]],
-  ['against_20tt_20xchange_20tt_1',['Linking your application against &lt;tt&gt;xchange&lt;/tt&gt;',['../index.html#autotoc_md12',1,'']]],
-  ['aggregate_20ids_2',['Aggregate IDs',['../index.html#autotoc_md20',1,'']]],
-  ['and_20elements_3',['Accessing substructures and elements',['../index.html#autotoc_md21',1,'']]],
-  ['and_20emitter_4',['JSON parser and emitter',['../index.html#autotoc_md26',1,'']]],
-  ['application_20against_20tt_20xchange_20tt_5',['Linking your application against &lt;tt&gt;xchange&lt;/tt&gt;',['../index.html#autotoc_md12',1,'']]],
-  ['arrays_6',['Arrays',['../index.html#autotoc_md18',1,'']]]
+  ['accessing_20substructures_20and_20elements_0',['Accessing substructures and elements',['../index.html#autotoc_md22',1,'']]],
+  ['added_1',['Added',['../md_CHANGELOG.html#autotoc_md3',1,'']]],
+  ['against_20tt_20xchange_20tt_2',['Linking your application against &lt;tt&gt;xchange&lt;/tt&gt;',['../index.html#autotoc_md13',1,'']]],
+  ['aggregate_20ids_3',['Aggregate IDs',['../index.html#autotoc_md21',1,'']]],
+  ['and_20elements_4',['Accessing substructures and elements',['../index.html#autotoc_md22',1,'']]],
+  ['and_20emitter_5',['JSON parser and emitter',['../index.html#autotoc_md27',1,'']]],
+  ['application_20against_20tt_20xchange_20tt_6',['Linking your application against &lt;tt&gt;xchange&lt;/tt&gt;',['../index.html#autotoc_md13',1,'']]],
+  ['arrays_7',['Arrays',['../index.html#autotoc_md19',1,'']]]
 ];