+++ apidoc/html/search/all_7.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['changelog_0',['Changelog',['../md_CHANGELOG.html',1,'']]],
-  ['changes_1',['Changes',['../md_CHANGELOG.html#autotoc_md3',1,'']]],
-  ['contents_2',['Table of Contents',['../index.html#autotoc_md6',1,'']]],
+  ['changed_0',['Changed',['../md_CHANGELOG.html#autotoc_md4',1,'']]],
+  ['changelog_1',['Changelog',['../md_CHANGELOG.html',1,'']]],
+  ['contents_2',['Table of Contents',['../index.html#autotoc_md7',1,'']]],
   ['contributing_20to_20the_20xchange_20library_3',['Contributing to the xchange library',['../md_CONTRIBUTING.html',1,'']]],
-  ['creating_20structure_4',['Creating structure',['../index.html#autotoc_md19',1,'']]]
+  ['creating_20structure_4',['Creating structure',['../index.html#autotoc_md20',1,'']]]
 ];