+++ apidoc/html/search/all_8.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['data_0',['Structured data',['../index.html#autotoc_md14',1,'']]],
-  ['data_20types_1',['Basic data types',['../index.html#autotoc_md15',1,'']]],
-  ['debugging_20support_2',['Debugging support',['../index.html#autotoc_md32',1,'']]]
+  ['data_0',['Structured data',['../index.html#autotoc_md15',1,'']]],
+  ['data_20types_1',['Basic data types',['../index.html#autotoc_md16',1,'']]],
+  ['debugging_20support_2',['Debugging support',['../index.html#autotoc_md33',1,'']]]
 ];