+++ apidoc/html/search/all_a.js
@@ -1,9 +1,9 @@
 var searchData=
 [
   ['false_0',['false',['../xjson_8c.html#aa93f0eb578d23995850d61f7d61c55c1',1,'FALSE:&#160;xjson.c'],['../xchange_8h.html#aa93f0eb578d23995850d61f7d61c55c1',1,'FALSE:&#160;xchange.h']]],
-  ['fields_1',['Sorting fields',['../index.html#autotoc_md24',1,'']]],
+  ['fields_1',['Sorting fields',['../index.html#autotoc_md25',1,'']]],
   ['firstfield_2',['firstField',['../structXStructure.html#a501ae8af2201d9cfaa60e3af21ff58fc',1,'XStructure']]],
-  ['fixes_3',['Fixes',['../md_CHANGELOG.html#autotoc_md2',1,'']]],
-  ['fragments_4',['JSON fragments',['../index.html#autotoc_md27',1,'']]],
-  ['future_20plans_5',['Future plans',['../index.html#autotoc_md33',1,'']]]
+  ['fixed_3',['Fixed',['../md_CHANGELOG.html#autotoc_md2',1,'']]],
+  ['fragments_4',['JSON fragments',['../index.html#autotoc_md28',1,'']]],
+  ['future_20plans_5',['Future plans',['../index.html#autotoc_md34',1,'']]]
 ];