+++ apidoc/html/xlookup_8c.js
@@ -4,6 +4,7 @@ var xlookup_8c =
     [ "xAllocLookup", "xlookup_8c.html#a01a94b3210aed984dbee956fbb26fcfd", null ],
     [ "xCreateLookup", "xlookup_8c.html#af8bee9eeac947d7e36ea436252e8dc2d", null ],
     [ "xDestroyLookup", "xlookup_8c.html#a02ea2d67e4c540ad706edcc1b3a0431e", null ],
+    [ "xDestroyLookupAndData", "xlookup_8c.html#a01f3e8af10b2ef8979eac5d7534648c1", null ],
     [ "xLookupCount", "xlookup_8c.html#a13e81b8f1bf3e1c3572ca74edb2836fe", null ],
     [ "xLookupField", "xlookup_8c.html#a5e3a316b7d0c5c40ed079f3df2d9fcc9", null ],
     [ "xLookupPut", "xlookup_8c.html#a3b423bf135d2ed42319abff99fda78e4", null ],