+++ doc/CHANGELOG.md
@@ -7,16 +7,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.1-rc1] - 2025-05-06
+## [Unreleased]
 
-### Fixes
+### Fixed
 
  - Handling of serialized strings in `xClearField()` and `xCopyOfField()`.
  
  - Handling of heterogeneous arrays (type `X_FIELD`) in `xCopyOfField()`.
+
+### Added
  
+ - `xDestroyLookupAndData()` to destroy a lookup table _including_ all the data that was referenced inside it. 
  
-### Changes
+### Changed
 
  - `xCreateField()` to treat `X_RAW` types always as scalars, ignoring the dimensions provided.